requirements.txt
unicornhat
unicornhatmini
flask
gpiozero
jsmin
discord
python-dotenv